refactor(language_server): refactor ignore code action logic as a linter fix (#14183)
part of https://github.com/oxc-project/oxc/issues/11707
`disable rule for this file` are only valid for the current `SubContextHost` and will produce syntax error for vue files.
A comment like this will be added.
```
// oxlint-disable no-debugger
<script>
debugger;
</script>
```

This is the first part to refactor the logic, by moving it nearer to the linter, where the source text / sections are available.
Probably needs some LSP feature flag for the message to pass the information of the current `ContextSubHost` source text offset.