refactor(linter/plugins): simplify creation of `context` in `defineRule` ESLint shim (#14206)
The `dummyOptions` and `dummyReport` default values are a useless complication. These values should never be accessed anyway - they're overridden before visitor methods access `context` - so `null` is fine.

I originally thought that storing objects of the right type here would help with making functions that touch the object monomorphic, but I now believe that to be mistaken. The property just has to be there to maintain consistent object shape, but what the *values* of the properties are doesn't matter.