refactor(linter/plugins): allow nullish values as `message` or `messageId` (#14422)
It makes the types a bit shit, but so far we accept `null` or `undefined` for all properties, and treat them the same as the property being absent. Extend this convention to `message` and `messageId` properties.

Also split out getting the message from a diagnostic into a separate function, to make this pattern simpler via early return.