refactor(linter/plugins): track ancestors while walking AST (#14432)
In the `oxlint` version of AST walker, track ancestors as traverse the AST.

This is a pre-requisite for supporting selectors DSL.

Ancestors are stored in an `ancestors` array which is exported by the module. Following what [esquery](https://www.npmjs.com/package/esquery) needs, the ancestors are stored in reverse order (i.e. `Program` is always the *last* in the array).