perf(linter): refactor rules to take advantage of node type skipping (#14519)
Updating rules to take advantage of the node type skipping so that they don't need to be run on every file. I forgot to split this out, so this is a composite PR of several rules, which I'll try to avoid in the future.

Some symbol-based rules have been refactored to search for the node type first and then look for symbols based on that. For example, if there are no imports in the file, we can skip `no-import-assign` instead of looping over all symbols and looking for an import.