feat(language_server): autodetect root `.oxfmtrc.json` (#14466)
> This PR implements autodetection of the root .oxfmtrc.json configuration file for the language server's formatter functionality. The language server can now automatically discover and use formatter configuration files located at the workspace root.

It does not support a custom config path, the next PR in this stack will implement it.
When a user does change the content of `.oxfmtrc.json`, he needs to restart the server at the moment.
Need to improve the file watcher implementation, created another stack with tests for current implementation.