refactor(linter): simplify built-in lint plugin checks (#14518)
Instead of checking that `plugin != "eslint"`, we can update `LintPlugins::from` to be `LintPlugins::try_from`, which allows for us to represent the default ESLint plugin. If it's any built-in plugin, this will return `Ok(LintPlugins::SomePlugin)`. If it's not a builtin, it will return `Err`. Previously, we had to check if it was empty and if it was not `eslint` plugin (since it was always empty by definition).