refactor(linter): remove `run_on_symbol` (#14610)
- closes https://github.com/oxc-project/backlog/issues/191

Previous PRs removed most usage of this run function, and this migrates the rest to use `run_once`. The motivation for this is not performance (this is likely a slight regression, actually), but should be a slight benefit to binary size and general maintenance. We now only have 3 ways of running lint rules, which makes the linter runtime code simpler.