fix(language_server): fix panic when "disable rule for this line" position is after error span (#14597)
The problem is, that `get_section_insert_position` get the offset after the line break.
But the error offset is before the insert offset. So it gets a range of `[4..3]`.

This rule makes no sense to ignore, because the span will always be 0 with a length of 0.
Ignored the case for adding "ignore fixes" and fixed the panic to make sure, it will not happen for other rules.

closes #14565