feat(ast_tools): support comment blocks in variants (#14681)
Add the ability to feature-gate code with comments as well as constants.

Before:

```js
if (THING) doStuff();
if (!THING) doOtherStuff();
```

After:

```js
/* IF THING */
doStuff();
/* END_IF */

/* IF !THING */
doOtherStuff();
/* END_IF */
```

The first form is still supported, but the latter is useful in places where the former doesn't work e.g.:

```js
/* IF THING */
export function doStuff() { /* ... */ }
/* END_IF */
```

If `THING` flag is not set for a variant, the code between the comments will be stripped out.