feat(language_server): watch for `fmt.configPath` file content change (#14509)
The server now sends a request to the client, to watch for `.oxfmtrc.json` or custom configurable `fmt.configPath` file content changes.
The client will send a `workspace/didChangeWatchedFiles` notification to the server.

The client can also send `workspace/didChangeConfiguration`, for now the server will restart both tools.
This will be optimized by looking at which tool is responsible for the file in https://github.com/oxc-project/oxc/pull/14645