perf(linter/plugins): lazy deserialize comments array (#14637)
Part of #14564.

Deserialize `program.comments` lazily upon first access, instead of
eagerly, on assumption that most JS plugins won't access comments.

`deserializeWith` cleans up the buffer once deserializing AST is
complete, so we keep a local copy of the buffer in scope of the
`comments` getter, and restore it when the getter is triggered.

Buffers are reused, so the `if (localAstId !== astId)` check is required
to make sure the AST in the buffer is same AST i.e. that the getter
isn't accessed *after* the file has completed linting.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: overlookmotel <theoverlookmotel@gmail.com>