refactor(language_server): improve file watching for different tools (#14645)
Now the worker is responsible to provide the dynamic registration with `Registration` and `Unregistration`.
The worker will use the provided tools to detect the needed watcher changes.
Before this PR, the server/Backend created the `Registration` struct for each worker, this was not great because we needed to register the watchers for the other tools too.
Now every tool has its own `Registration.id` which can be unregistered.

https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#client_registerCapability
https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#client_unregisterCapability