perf(linter): `id-length`: check if ident is ASCII before segmenting (#14767)
Profiling shows that segmenting into graphemes is much slower than just looking at the number of bytes. A simple check on `is_ascii` allows us to avoid splitting into graphemes and use the identifier length instead, which is much faster. We could optimize this further in the future by doing some faster ASCII check, but this is a good start.