perf(linter): `no-dupe-class-members`: only run when there are any classes (#14867)
Similar to https://github.com/oxc-project/oxc/pull/14865. Allows removing this rule from the array of rules if there are no classes in the file.

Benchmarks don't really show any difference here, but it might just be because we haven't yet crossed an allocation threshold (i.e., once we remove enough rules, we'll see a diff)