fix(formatter): `BestFitting` doesn't exactly matches the `conditinalGroup` behavior in Prettier (#14983)
I found a case where Oxfmt's output differs from Prettier. Honestly, I think the `Oxfmt` output is prettier than `Prettier`.
Anyway, I investigated this case to find the root cause.

### Failed example

Input:
```ts
const defaultColorDecoratorsEnablement = accessor
  .get(IConfigurationService)
  .getValue<"auto" | "always" | "never">(
    "longlonglonglonglonglonglonglonglong",
  );
```

Oxfmt and Biome's output:
```ts
const defaultColorDecoratorsEnablement = accessor
  .get(IConfigurationService)
  .getValue<"auto" | "always" | "never">(
    "longlonglonglonglonglonglonglonglong",
  );
```

Prettier output:
```ts
const defaultColorDecoratorsEnablement = accessor
  .get(IConfigurationService)
  .getValue<
    "auto" | "always" | "never"
  >("longlonglonglonglonglonglonglonglong");
```
### Investigating the cause

Initially, I think the problem is that we generate inconsistent `IR`, so the output is different, which is a reasonable suspicion because, so far, almost all issues are due to this reason. However, after carefully comparing both IR of Oxfmt and Prettier, they are essentially the same, so the problem is moved to the printing IR part.

After spending a few days on this case, I finally found the reason why they produce the same IR but print different outputs. The cause is inside the `BestFitting`, which is a graceful solution for Prettier's `conditionalGroup`.

Let's take a look at how Prettier handles `conditionalGroup`

#### Prettier implementation:
**conditionalGroup**
https://github.com/prettier/prettier/blob/7584432401a47a26943dd7a9ca9a8e032ead7285/src/document/builders.js#L99-L106
```js
/**
 * @param {Doc[]} states
 * @param {object} [opts] - TBD ???
 * @returns Doc
 */
function conditionalGroup(states, opts) {
  return group(states[0], { ...opts, expandedStates: states });
}
```

**handling of conditionalGroup**
https://github.com/prettier/prettier/blob/7584432401a47a26943dd7a9ca9a8e032ead7285/src/document/printer.js#L408-L445
```js
// Expanded states are a rare case where a document
// can manually provide multiple representations of
// itself. It provides an array of documents
// going from the least expanded (most flattened)
// representation first to the most expanded. If a
// group has these, we need to manually go through
// these states and find the first one that fits.
// eslint-disable-next-line no-lonely-if
if (doc.expandedStates) {
  const mostExpanded = doc.expandedStates.at(-1);

  if (doc.break) {
    cmds.push({ ind, mode: MODE_BREAK, doc: mostExpanded });

    break;
  } else {
    for (let i = 1; i < doc.expandedStates.length + 1; i++) {
      if (i >= doc.expandedStates.length) {
        cmds.push({ ind, mode: MODE_BREAK, doc: mostExpanded });

        break;
      } else {
        const state = doc.expandedStates[i];
        /** @type {Command} */
        const cmd = { ind, mode: MODE_FLAT, doc: state };

        if (fits(cmd, cmds, rem, hasLineSuffix, groupModeMap)) {
          cmds.push(cmd);

          break;
        }
      }
    }
  }
} else {
  cmds.push({ ind, mode: MODE_BREAK, doc: doc.contents });
}
```

#### Oxfmt implementation

https://github.com/oxc-project/oxc/blob/d9f7df8bc6bf7e48c90ba99638502c87afba0e04/crates/oxc_formatter/src/formatter/printer/mod.rs#L366-L418

The difference between the implementations is in handling different `expandedStates` (Called as `variants` in Oxfmt).

In prettier, they will try to get all `expandedStates` except the first one to check whether they fit or not in the `MODE_FLAT` mode. If not, they will fall back to the `mostExpanded` and print it in `MODE_BREAK` mode.