docs(linter): Add configuration option docs for jest/require-top-level-describe rule. (#14973)
These already existed, but this ensures they're done via the
documentation auto-generation tooling.

Part of #14743.

Generated docs:

```md
## Configuration

This rule accepts a configuration object with the following properties:

### maxNumberOfTopLevelDescribes

type: `integer`

default: `18446744073709551615`

The maximum number of top-level `describe` blocks allowed in a test file.
```

This one is a bit weird in that the default is the MAX usize value, but
that's probably fine to expose?