docs(linter): Add configuration option docs for eslint/no-labels rule. (#15112)
Part of #14743.

Generated docs:

```md
## Configuration

This rule accepts a configuration object with the following properties:

### allowLoop

type: `boolean`

default: `false`

If set to `true`, this rule ignores labels which are sticking to loop statements.
Examples of **correct** code with this option set to `true`:
\```js
label:
while (true) {
break label;
}
\```


### allowSwitch

type: `boolean`

default: `false`

If set to `true`, this rule ignores labels which are sticking to switch statements.
Examples of **correct** code with this option set to `true`:
\```js
label:
switch (a) {
case 0:
break label;
}
\```
```