docs(linter): Add configuration option docs for typescript/no-empty-interface rule. (#15098)
Part of #14743.

As an aside, this rule [is deprecated in the TypeScript ESLint
plugin](https://typescript-eslint.io/rules/no-empty-interface/), maybe
it should be deprecated in oxlint as well? Or at least mention that the
`typescriptno-empty-object-type` rule is preferred over it?

Generated docs:

```md
## Configuration

This rule accepts a configuration object with the following properties:

### allowSingleExtends

type: `boolean`

default: `false`

When set to `true`, allows empty interfaces that extend a single interface.
```