docs(linter): Add configuration option docs for unicorn/prefer-number-properties rule. (#15141)
Part of #14743.

Generated docs:

```md
## Configuration

This rule accepts a configuration object with the following properties:

### checkInfinity

type: `boolean`

default: `false`

If set to `true`, checks for usage of `Infinity` and `-Infinity` as global variables.


### checkNan

type: `boolean`

default: `true`

If set to `true`, checks for usage of `NaN` as a global variable.
```