docs(linter): Add config option docs for 3 rules. (#15211)
Part of #14743.

- eslint/no-bitwise
- eslint/no-console
- eslint/no-inner-declarations

Generated docs:

```md
## Configuration

This rule accepts a configuration object with the following properties:

### allow

type: `string[]`

default: `[]`

The `allow` option permits the given list of console methods to be used as exceptions to
this rule.

Say the option was configured as `{ "allow": ["info"] }` then the rule would behave as
follows:

Example of **incorrect** code for this option:
\```javascript
console.log('foo');
\```

Example of **correct** code for this option:
\```javascript
console.info('foo');
\```
```

```md
## Configuration

This rule accepts a configuration object with the following properties:

### blockScopedFunctions


default: `null`

Controls whether function declarations in nested blocks are allowed in strict mode (ES6+ behavior).


### config


default: `"functions"`

Determines what type of declarations to check.
```

```md
## Configuration

This rule accepts a configuration object with the following properties:

### allow

type: `string[]`

default: `[]`

The `allow` option permits the given list of bitwise operators to be used
as exceptions to this rule.

For example `{ "allow": ["~"] }` would allow the use of the bitwise operator
`~` without restriction. Such as in the following:

\```javascript
~[1,2,3].indexOf(1) === -1;
\```


### int32Hint

type: `boolean`

default: `false`

When set to true the `int32Hint` option allows the use of bitwise OR in |0
pattern for type casting.

For example with `{ "int32Hint": true }` the following is permitted:

\```javascript
const b = a|0;
\```
```